fix: Use float for health/damage, int for shields throughout codebase
Fixed type consistency across the codebase to match domain model:
- Health and damage values: float (supports fractional damage like 2.5)
- Shield values: int (discrete shield counts)

## Domain (already correct)
- Player.Health, Player.MaxHealth: float
- Player.Shields: int
- Player.TakeDamage(float) returns float
- Player.Heal(float) returns float

## DTOs Updated
- PlayerDTO.Health: int → float
- PlayerDTO.MaxHealth: int → float

## Events Updated
- PlayerDamagedEvent: Damage, HealthDamage, CurrentHealth → float
- PlayerHealedEvent: HealAmount, NewHP → float
- PlayerHealthChangedEvent: OldHealth, NewHealth, Delta → float

## Use Cases & Abilities Updated
- DamagePlayerAbility: healthDamage variable → float
- HealPlayerAbility: actualHealed variable → float
- AttackUseCase: healthDamage variable → float
- AttackResult.HealthDamage: int → float
- AttackResult.DirectAttackSuccess signature updated

## Services Fixed
- Removed readonly from _modernAbilityDictionary fields in:
  - FieldAbilityProviderService
  - EquipmentAbilityProviderService
  - EffectAbilityProviderService
  (readonly fields cannot be assigned in methods, even from constructor)

## Tests Updated
- GameFlowIntegrationTests: Updated health variables to float

+++ JDG Mobile Game/Assets/Tests/JDG.Infrastructure.Tests/Integration/GameFlowIntegrationTests.cs
@@ -184,11 +184,11 @@ namespace JDG.Infrastructure.Tests.Integration
             var player = _playerRepository.CreatePlayer(PlayerId.Player1, new[] { card.Id }, maxHealth: 30);
             Assert.IsNotNull(player, "Player should be created");
 
-            int initialHealth = player.Health;
+            float initialHealth = player.Health;
 
             // Act: Direct damage through domain entity
-            int damage = 10;
-            int actualDamage = player.TakeDamage(damage);
+            float damage = 10;
+            float actualDamage = player.TakeDamage(damage);
             _playerRepository.SavePlayer(player);
 
             // Assert
@@ -255,7 +255,7 @@ namespace JDG.Infrastructure.Tests.Integration
 
             // Act: Direct attack on Player 2
             var player2Before = _playerRepository.GetPlayer(PlayerId.Player2);
-            int healthBefore = player2Before.Health;
+            float healthBefore = player2Before.Health;
 
             var result = attackUseCase.ExecuteDirectAttack(PlayerId.Player1, cardToPlay.Id, PlayerId.Player2);
